requirements.txt
torch
torchvision
numpy
pillow
tqdm
